backend/requirements.txt
Flask==2.3.3
peewee==3.16.3
python-dotenv==1.0.0
PyJWT==2.8.0
waitress==2.1.2
Flask-Cors==4.0.0
bcrypt==4.0.1
Werkzeug==2.3.7 